requirements.txt
fastapi
uvicorn
pandas
scikit-learn
joblib
streamlit
requests
sentence-transformers
transformers
torch
numpy
google-generativeai
